% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefs.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Posterior predictive check for walker object}
\usage{
plot_coefs(object, level = 0.05, alpha = 0.33)
}
\arguments{
\item{object}{An output from \code{\link{walker}}.}

\item{level}{Level for intervals. Default is 0.05, leading to 90\% intervals.}

\item{alpha}{Transparency level for \code{geom_ribbon}.}
}
\description{
Plots sample quantiles from posterior predictive sample. 
See \code{\link[bayesplot]{ppc_ribbon}} for details.
}
