% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot_predict}
\alias{plot_predict}
\title{Prediction intervals for walker object}
\usage{
plot_predict(object, level = 0.05, alpha = 0.33)
}
\arguments{
\item{object}{An output from \code{\link{predict.walker_fit}}.}

\item{level}{Level for intervals. Default is 0.05, leading to 90\% intervals.}

\item{alpha}{Transparency level for \code{geom_ribbon}.}
}
\description{
Plots sample quantiles from pre
See \code{\link[bayesplot]{ppc_ribbon}} for details.
}
