% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit.R
\name{print.walker_fit}
\alias{print.walker_fit}
\title{Print Summary of walker_fit Object}
\usage{
\method{print}{walker_fit}(object, level = 0.05, alpha = 0.33)
}
\arguments{
\item{object}{An output from \code{\link{predict.walker_fit}}.}

\item{level}{Level for intervals. Default is 0.05, leading to 90\% intervals.}

\item{alpha}{Transparency level for \code{\link{geom_ribbon}}.}
}
\description{
Plots sample quantiles from pre
See \code{\link{ppc_ribbon}} for details.
}
