% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{walker_rw2}
\alias{walker_rw2}
\title{Bayesian linear regression with second order random walk coefficients}
\usage{
walker_rw2(formula, data, beta_prior, sigma_prior, init, chains, newdata,
  naive = FALSE, return_x_reg = FALSE, return_y_rep = TRUE, ...)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}. See \code{\link[stats]{lm}} for details.}

\item{data}{An optional data.frame or object coercible to such, as in \code{\link[stats]{lm}}.}

\item{beta_prior}{A matrix with \eqn{k} rows and 2 columns, where first columns defines the 
prior means of the Gaussian priors of the corresponding \eqn{k} regression coefficients, 
and the second column defines the the standard deviations of those prior distributions.}

\item{sigma_prior}{A matrix with \eqn{k + 1} rows and two colums with similar structure as 
\code{beta_prior}, with first row corresponding to the prior of the standard deviation of the 
observation level noise, and rest of the rows define the priors for the standard deviations of 
random walk noise terms. The prior distributions for all sigmas are 
Gaussians truncated to positive real axis. For non-Gaussian models, this should contain only k rows. 
For second order random walk model, these priors correspond to the slope level standard deviations.}

\item{init}{Initial value specification, see \code{\link[rstan]{sampling}}. 
Note that compared to default in \code{rstan}, here the default is a to sample from the priors.}

\item{chains}{Number of Markov chains. Default is 4.}

\item{newdata}{Optional data.frame containing covariates used for prediction. This argument is 
ignored if argument \code{naive} is \code{TRUE}.}

\item{naive}{Only used for \code{walker} function. 
If \code{TRUE}, use "standard" approach which samples the joint posterior 
\eqn{p(beta, sigma | y)}. If \code{FALSE} (the default), use marginalisation approach 
where we sample the marginal posterior \eqn{p(sigma | y)} and generate the samples of 
\eqn{p(beta | sigma, y)} using state space modelling techniques 
(namely simulation smoother by Durbin and Koopman (2002)). Both methods give asymptotically 
identical results, but the latter approach is computationally much more efficient.}

\item{return_x_reg}{If \code{TRUE}, does not perform sampling, but instead returns the matrix of 
predictors after processing the \code{formula}.}

\item{return_y_rep}{If \code{TRUE} (default), \code{walker} also returns the samples from the 
posterior predictive distribution \eqn{p(y_rep | y)}. This argument is ignored if 
argument \code{naive} is \code{TRUE}.}

\item{...}{Further arguments to \code{\link[rstan]{sampling}}.}
}
\description{
Function \code{walker_rw2} replaces the first order random walk coefficients 
with smooth second order random walks.
}
\seealso{
walker walker_glm
}
