% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit.R
\name{as.data.frame.walker_fit}
\alias{as.data.frame.walker_fit}
\title{Coerce Posterior Samples of walker Fit to a Data Frame}
\usage{
\method{as.data.frame}{walker_fit}(x, type, ...)
}
\arguments{
\item{x}{An output from \code{\link{walker}} or \code{\link{walker_glm}}.}

\item{type}{Either \code{tiv} (time-invariant parameters) or \code{tv} (time-varying coefficients).}
}
\description{
Creates a data.frame object from the output of walker fit.
}
