% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.walker_fit}
\alias{fitted.walker_fit}
\title{Extract Fitted Values of Walker Fit}
\usage{
\method{fitted}{walker_fit}(object, summary = TRUE, ...)
}
\arguments{
\item{object}{Output of \code{walker} or \code{walker_glm}.}

\item{summary}{If \code{TRUE} (default), return summary statistics. Otherwise returns samples.}

\item{...}{Ignored.}
}
\value{
Time series containing fitted values.
}
\description{
Returns fitted values (posterior means) from output of \code{walker} or \code{walker_glm}.
}
