% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefs.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Posterior predictive check for walker object}
\usage{
plot_coefs(
  object,
  level = 0.05,
  alpha = 0.33,
  transform = identity,
  scales = "fixed",
  add_zero = TRUE
)
}
\arguments{
\item{object}{An output from \code{\link{walker}}.}

\item{level}{Level for intervals. Default is 0.05, leading to 90\% intervals.}

\item{alpha}{Transparency level for \code{geom_ribbon}.}

\item{transform}{Optional vectorized function for transforming the coefficients (for example \code{exp}).}

\item{scales}{Should y-axis of the panels be \code{"fixed"} (default) or \code{"free"}?}

\item{add_zero}{Logical, should a dashed line indicating a zero be included?}
}
\description{
Plots sample quantiles from posterior predictive sample. 
See \code{\link{ppc_ribbon}} for details.
}
