% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check.walker_fit}
\alias{pp_check.walker_fit}
\title{Posterior predictive check for walker object}
\usage{
\method{pp_check}{walker_fit}(object, ...)
}
\arguments{
\item{object}{An output from \code{\link{walker}}.}

\item{...}{Further parameters to \code{\link{ppc_ribbon}}.}
}
\description{
Plots sample quantiles from posterior predictive sample. 
See \code{\link{ppc_ribbon}} for details.
}
