% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit.R
\name{print.walker_fit}
\alias{print.walker_fit}
\title{Print Summary of walker_fit Object}
\usage{
\method{print}{walker_fit}(x, ...)
}
\arguments{
\item{x}{An output from \code{\link[=walker]{walker()}} or \code{\link[=walker_glm]{walker_glm()}}.}

\item{...}{Additional arguments to \code{\link[rstan:print.stanfit]{rstan::print.stanfit()}}.}
}
\description{
Prints the summary information of time-invariant model parameters. In case of non-Gaussian models,
results based on approximate model are returned with a warning.
}
