% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw2}
\alias{rw2}
\title{Construct a first-order random walk component}
\usage{
rw2(formula, data, beta_prior, sigma_prior, slope_prior)
}
\arguments{
\item{formula}{Formula for RW1 part of the model. Only right-hand-side is used.}

\item{data}{Optional data.frame.}

\item{beta_prior}{A vector of length two which defines the 
prior mean and standard deviation of the Gaussian prior for coefficients at time 1.}

\item{sigma_prior}{A vector of length two, defining the truncated Gaussian prior for 
the slope level standard deviation.}

\item{slope_prior}{A vector of length two which defines the 
prior mean and standard deviation of the Gaussian prior for the slopes at time 1.}
}
\description{
Auxiliary function used inside of the formula of \code{walker}.
}
