% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fit.R
\name{summary.walker_fit}
\alias{summary.walker_fit}
\title{Summary of walker_fit Object}
\usage{
\method{summary}{walker_fit}(object, type = "tiv", ...)
}
\arguments{
\item{object}{An output from \code{\link{walker}} or \code{\link{walker_glm}}.}

\item{type}{Either \code{tiv} (time-invariant parameters, the default) or \code{tv} (time-varying coefficients).}

\item{...}{Ignored.}
}
\description{
Return summary information of time-invariant model parameters.
}
